MAINT: syntax updates and fixes for Python 3.9 and pandas 1.3

+++ zipline/protocol.py
@@ -174,7 +174,7 @@ class Portfolio(object):
                     asset.price_multiplier
             )
             for asset, position in self.positions.items()
-        })
+        }, dtype="float64")
         return position_values / self.portfolio_value
